Use "_x64" instead of "amd64" for x64 MSW makefile builds.
Still accept "amd64" as a valid TARGET_CPU value but also accept, and
encourage people to use, "x64".

Add a new section about the build system changes to the change log and added a
mini-TOC to it to make navigating it a bit easier and show what changes are
there at a glance.